requirements.txt
absl-py==2.3.1
accelerate==1.11.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
anywidget==0.9.21
asttokens==3.0.1
astunparse==1.6.3
asyncpg==0.30.0
bbox_visualizer==0.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Cython==3.2.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
einops==0.8.1
email-validator==2.3.0
executing==2.2.1
fairscale==0.4.13
fastapi==0.121.2
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
ftfy==6.3.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.7
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_bbox_widget==0.6.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
limits==5.6.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
nest_asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnx==1.19.1
onnx_graphsurgeon==0.5.8
onnx-ir==0.1.12
onnx-tf==1.10.0
onnxruntime==1.23.2
onnxscript==0.5.6
onnxsim==0.4.36
open_clip_torch==3.2.0
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
passlib==1.7.4
peft==0.18.0
pexpect==4.9.0
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
pip==25.3
platformdirs==4.5.0
polygraphy==0.49.26
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
psycopg2-binary==2.9.11
psygnal==0.15.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycocotools==2.0.10
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydantic-settings==2.12.0
Pygments==2.19.2
pylabel==0.1.55
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rf100vl==1.1.0
rfdetr==1.3.0
rich==14.2.0
roboflow==1.2.11
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.44
stack_data==0.6.3
starlette==0.49.3
supervision==0.27.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-addons==0.23.0
termcolor==3.2.0
threadpoolctl==3.6.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.5.1
typeguard==2.13.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.15
wrapt==2.0.1
zipp==3.23.0
